third_party/blink/web_tests/external/wpt/content-security-policy/reporting/report-clips-sample.https-expected.txt

This is a testharness.js-based test.
[FAIL] Function constructor - the other kind of eval - is clipped.
  assert_equals: expected "Function|(a,b\\n) {\\nreturn return '1234567890123456" but got "Function|(a,b\\n) {\\nreturn '12345678901234567890123"
[FAIL] Async Function constructor is also clipped.
  assert_equals: expected "Function|(a,b\\n) {\\nreturn return '1234567890123456" but got "eval|(async function anonymous(a,b\\n) {\\nreturn"
[FAIL] Generator Function constructor is also clipped.
  assert_equals: expected "Function|(a,b\\n) {\\nreturn return '1234567890123456" but got "eval|(function* anonymous(a,b\\n) {\\nreturn '123"
[FAIL] AsyncGenerator Function constructor is also clipped.
  assert_equals: expected "Function|(a,b\\n) {\\nreturn return '1234567890123456" but got "eval|(async function* anonymous(a,b\\n) {\\nretur"
Harness: the test ran to completion.
